Move v1 process output resolution to TaskOutputResolverV1
